refactor(fetch): move fetch polyfill to file
