+++ tsconfig.json
@@ -44,7 +44,7 @@
     "paths": {
       /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
       "src/lib/keychain": ["./src/polyfills/keychain"],
-      "rn-fetch-blob": ["./src/node.polyfill"],
+      "rn-fetch-blob": ["./src/polyfills/rnFetchBlob"],
       "src/lib/errors/sentry": ["./src/node.polyfill"],
       "react-native-localize": ["./src/node.polyfill"],
       "src/locales/i18n": ["./src/node.polyfill"],
